% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplot.R
\name{kplot}
\alias{kplot}
\alias{kplot.default}
\title{kplot Function}
\usage{
kplot(x, ...)

\method{kplot}{default}(x, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{...}{Additional arguments to be passed to the plotting method.}
}
\value{
A plot of the object \code{x}.
}
\description{
This function is a generic method for plotting objects. It uses the \code{UseMethod} function to dispatch the appropriate method based on the class of the object \code{x}.
}
\examples{
kplot(1:10)

}
